+++ ManaKit/Classes/DeckTableViewCell.swift
@@ -57,7 +57,9 @@ public class DeckTableViewCell: UITableViewCell {
     
         // thumbnail image
         if let card = deck.heroCard {
-            if let croppedImage = ManaKit.sharedInstance.croppedImage(card) {
+            if let croppedImage = ManaKit.sharedInstance.cardImage(card,
+                                                                   imageType: .artCrop,
+                                                                   roundCornered: false) {
                 thumbnailImage.image = croppedImage
             } else {
                 thumbnailImage.image = ManaKit.sharedInstance.imageFromFramework(imageName: .cardBackCropped)
@@ -65,7 +67,9 @@ public class DeckTableViewCell: UITableViewCell {
                 firstly {
                     ManaKit.sharedInstance.downloadImage(ofCard: card, imageType: .artCrop)
                 }.done {
-                    guard let image = ManaKit.sharedInstance.croppedImage(card) else {
+                    guard let image = ManaKit.sharedInstance.cardImage(card,
+                                                                       imageType: .artCrop,
+                                                                       roundCornered: false) else {
                         return
                     }